// libs/human-resource/src/lib/models/employees.ts
export interface EmployeeData {
  cif: string;
  fullName: string;
  dateOfBirth: string;
  gender: string;
  maritalStatus: string;
  phoneNumber: string;
  team: string;
  email: string;
  skype: string;
  office: string;
  nationality: string;
  languages: string[];
  jobTitle: string;
  joinDate: string;
  major: string[] | string;
  institute: string[] | string;
  facebook: string;
  instagram: string;
}